Implement default values for source definitions.
This commit implements the entirety of default parameters for queries.

During normal resolution a query is matched againsnt source definitions
until a valid source definition is found (we circuit-break to the first
valid).

If the query hasn't specified parameters with defaults for the valid
source defintiion the query will be supplemented with the defaults from
the source definition.

The supplemented query's hash from .toString() will be stored in the
cache, therefore each resolution now has two cache lookups:  once for
the user-generated query and once for the supplemented query with
defaults.  This happens prior to calling the source definition for a
query so we can use the cache for existing data.

Ensure that a query satisfies a source when it omits required params
with defaults